refactoring the composable functions
