+++ app/src/main/java/com/gathering/android/auth/AuthComposables.kt
@@ -1,0 +1,74 @@
+package com.gathering.android.auth
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+@Preview
+fun AuthButtonPreview() {
+    AuthButton(text = "Button", onClick = { /*TODO*/ })
+}
+
+@Composable
+fun AuthButton(
+    text: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    isLoading: Boolean = false
+) {
+    Button(
+        shape = RoundedCornerShape(0.dp),
+        onClick = onClick,
+        modifier = modifier
+            .fillMaxWidth()
+            .height(50.dp),
+        colors = ButtonDefaults.buttonColors(
+            containerColor = Color.DarkGray
+        )
+    ) {
+        if (isLoading) {
+            CircularProgressIndicator(color = Color.White)
+        } else {
+            Text(text)
+        }
+    }
+}
+
+@Composable
+fun CustomUnderlinedButton(text: String, onClick: () -> Unit) {
+    val underlinedText = remember(text) {
+        AnnotatedString.Builder(text).apply {
+            addStyle(
+                style = SpanStyle(textDecoration = TextDecoration.Underline),
+                start = 0,
+                end = text.length
+            )
+        }.toAnnotatedString()
+    }
+    Button(
+        onClick = onClick,
+        colors = ButtonDefaults.buttonColors(
+            containerColor = Color.Transparent,
+            contentColor = Color.Black,
+        )
+    ) {
+        Text(
+            text = underlinedText
+        )
+    }
+}
