+++ app/src/main/java/com/gathering/android/common/Composables.kt
@@ -1,18 +1,15 @@
 package com.gathering.android.common
 
 import androidx.annotation.StringRes
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
@@ -21,7 +18,6 @@ import androidx.compose.material.icons.filled.VisibilityOff
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
@@ -33,21 +29,14 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.CornerRadius
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.gathering.android.R
@@ -138,76 +127,30 @@ fun CustomActionButton(
         containerColor = Color.DarkGray
     )
 ) {
-    Button(shape = RoundedCornerShape(0.dp),
-        onClick = onClick,
-        modifier = modifier
-            .height(60.dp)
-            .width(170.dp),
-        colors = colors,
-        contentPadding = PaddingValues(8.dp),
-        content = {
-            Text(
-                text = text,
-                modifier = Modifier.padding(4.dp),
-            )
-        })
-    if (isLoading) {
-        CircularProgressIndicator(color = Color.White)
-    } else {
-        Text(text)
-    }
-}
-
-@Composable
-fun CustomUnderlinedButton(text: String, onClick: () -> Unit) {
-    val underlinedText = remember(text) {
-        AnnotatedString.Builder(text).apply {
-            addStyle(
-                style = SpanStyle(textDecoration = TextDecoration.Underline),
-                start = 0,
-                end = text.length
-            )
-        }.toAnnotatedString()
-    }
-    Button(
-        onClick = onClick,
-        colors = ButtonDefaults.buttonColors(
-            containerColor = Color.Transparent,
-            contentColor = Color.Black,
-        )
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(10.dp),
+        horizontalArrangement = Arrangement.Center
     ) {
-        Text(
-            text = underlinedText
+        Button(
+            shape = RoundedCornerShape(0.dp),
+            onClick = onClick,
+            modifier = modifier
+                .height(60.dp)
+                .width(170.dp),
+            colors = colors,
+            contentPadding = PaddingValues(8.dp),
+            content = {
+                Text(
+                    text = text,
+                    modifier = Modifier.padding(4.dp),
+                )
+            }
         )
     }
-}
 
-@Composable
-@Preview
-fun AuthButtonPreview() {
-    AuthButton(text = "Button", onClick = { /*TODO*/ })
 }
-
-@Composable
-fun AuthButton(
-    text: String, onClick: () -> Unit, modifier: Modifier = Modifier, isLoading: Boolean = false
-) {
-    Button(
-        shape = RoundedCornerShape(0.dp),
-        onClick = onClick,
-        modifier = modifier.fillMaxWidth(),
-        colors = ButtonDefaults.buttonColors(
-            containerColor = Color.DarkGray
-        )
-    ) {
-        if (isLoading) {
-            CircularProgressIndicator(color = Color.White)
-        } else {
-            Text(text)
-        }
-    }
-}
-
 @Composable
 @Preview(showBackground = true)
 fun CustomTextViewPreview() {
@@ -221,7 +164,9 @@ fun CustomTextView(
     textStyle: TextStyle = TextStyle.Default
 ) {
     Box(
-        modifier = modifier.fillMaxWidth()
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(10.dp)
     ) {
         val text = stringResource(id = textResId)
         Text(
@@ -244,66 +189,12 @@ fun ErrorText(error: String) {
         text = error,
         modifier = Modifier
             .fillMaxWidth()
-            .padding(10.dp),
+            .padding(25.dp),
         textAlign = TextAlign.Center,
         style = TextStyle(
             fontSize = 17.sp, color = Color.Red
         )
     )
-}
-
-@Composable
-fun PageIndicatorView(
-    isSelected: Boolean,
-    selectedColor: Color,
-    defaultColor: Color,
-    defaultRadius: Dp,
-    selectedLength: Dp,
-    animationDurationInMillis: Int,
-    modifier: Modifier = Modifier,
-) {
-
-    val color: Color by animateColorAsState(
-        targetValue = if (isSelected) {
-            selectedColor
-        } else {
-            defaultColor
-        },
-        animationSpec = tween(
-            durationMillis = animationDurationInMillis,
-        ), label = ""
-    )
-    val width: Dp by animateDpAsState(
-        targetValue = if (isSelected) {
-            selectedLength
-        } else {
-            defaultRadius
-        },
-        animationSpec = tween(
-            durationMillis = animationDurationInMillis,
-        ), label = ""
-    )
-
-    Canvas(
-        modifier = modifier
-            .size(
-                width = width,
-                height = defaultRadius,
-            ),
-    ) {
-        drawRoundRect(
-            color = color,
-            topLeft = Offset.Zero,
-            size = Size(
-                width = width.toPx(),
-                height = defaultRadius.toPx(),
-            ),
-            cornerRadius = CornerRadius(
-                x = defaultRadius.toPx(),
-                y = defaultRadius.toPx(),
-            ),
-        )
-    }
 }
