+++ app/src/main/java/com/gathering/android/intro/IntroComposables.kt
@@ -1,22 +1,31 @@
 package com.gathering.android.intro
 
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.CornerRadius
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.gathering.android.common.PageIndicatorView
 
 
 @Composable
 @Preview
-fun pageIndicatorPreview() {
+fun PageIndicatorPreview() {
     PageIndicatorView(
         isSelected = true,
         selectedColor = Color.Red,
@@ -51,5 +60,59 @@ fun PageIndicator(
                 modifier = Modifier.padding(8.dp)
             )
         }
+    }
+}
+
+@Composable
+fun PageIndicatorView(
+    isSelected: Boolean,
+    selectedColor: Color,
+    defaultColor: Color,
+    defaultRadius: Dp,
+    selectedLength: Dp,
+    animationDurationInMillis: Int,
+    modifier: Modifier = Modifier,
+) {
+
+    val color: Color by animateColorAsState(
+        targetValue = if (isSelected) {
+            selectedColor
+        } else {
+            defaultColor
+        },
+        animationSpec = tween(
+            durationMillis = animationDurationInMillis,
+        ), label = ""
+    )
+    val width: Dp by animateDpAsState(
+        targetValue = if (isSelected) {
+            selectedLength
+        } else {
+            defaultRadius
+        },
+        animationSpec = tween(
+            durationMillis = animationDurationInMillis,
+        ), label = ""
+    )
+
+    Canvas(
+        modifier = modifier
+            .size(
+                width = width,
+                height = defaultRadius,
+            ),
+    ) {
+        drawRoundRect(
+            color = color,
+            topLeft = Offset.Zero,
+            size = Size(
+                width = width.toPx(),
+                height = defaultRadius.toPx(),
+            ),
+            cornerRadius = CornerRadius(
+                x = defaultRadius.toPx(),
+                y = defaultRadius.toPx(),
+            ),
+        )
     }
 }
